+++ packages/primitives/src/components/Tab/TabPanels.tsx
@@ -1,0 +1,18 @@
+import { Children, cloneElement } from 'react';
+
+import type { TabPanelsProps } from './tab.types';
+
+const TabPanels = (props: TabPanelsProps) => {
+  const { children } = props;
+  return (
+    <>
+      {Children.map(children, (child, index) =>
+        cloneElement(child, {
+          tabId: index,
+        }),
+      )}
+    </>
+  );
+};
+
+export default TabPanels;
